libm: Add optimized powf
This implementation is based on code from ARM optimized routines [1].
The error for powf, for nearest rouding mode, is estimated on 0.82 ulp
with worst case.

The function is implemented as exp2(y*log2(x)) with the same algorithms
used in exp2f and log2f, except that the log2f polynomial is larger for
extra precision and its output (and exp2f input) may be scaled by a
power of 2 (POWF_SCALE) to simplify the argument reduction step of exp2
(possible when efficient round and convert toint operation is
available).

The special case handling tries to minimize the checks in the hot path.
When the input of exp2_inline is checked, int arithmetics is used as
that was faster on the tested aarch64 cores.

The performance shows about 5.3x better throughput and 3.6x better
latency for aarch64 (A72) using powf benchmark.

[1] https://github.com/ARM-software/optimized-routines

Test: ran 32-bit and 64-bit x86 tests on host
Change-Id: Id8ed5383e05dec781e187d52e60d41e83f87f1cb
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>